# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for EVE Online trading, manufacturing and profit optimization\n\nFeatures:\n- Intra-region trading route calculation\n- Character skills and ship fitting integration\n- Deterministic cargo/warp speed calculations\n- Market data caching with staleness tracking\n- EVE SSO authentication",
        "title": "EVE-O-Provit API",
        "contact": {
            "name": "EVE-O-Provit Team",
            "url": "https://github.com/Sternrassler/eve-o-provit",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/market/staleness/{region}": {
            "get": {
                "description": "Check how old the cached market data is for a region\nStatus: fresh (\u003c30min), stale (30-60min), very_stale (\u003e60min)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Market"
                ],
                "summary": "Get market data staleness",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 10000002,
                        "description": "Region ID",
                        "name": "region",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MarketDataStalenessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/market/{region}/{type}": {
            "get": {
                "description": "Retrieve market orders for a specific item type in a region\nSupports optional refresh from ESI (cached for 5 minutes)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Market"
                ],
                "summary": "Get market orders",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 10000002,
                        "description": "Region ID",
                        "name": "region",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 34,
                        "description": "Type ID",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Force refresh from ESI",
                        "name": "refresh",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/MarketOrderResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sde/regions": {
            "get": {
                "description": "Get list of all EVE Online regions from SDE",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SDE"
                ],
                "summary": "List all EVE regions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/RegionResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/types/{id}": {
            "get": {
                "description": "Retrieve detailed information about an EVE Online item type from SDE",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SDE"
                ],
                "summary": "Get item type information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TypeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                },
                "message": {
                    "type": "string",
                    "example": "Detailed error message"
                }
            }
        },
...
        "MarketDataStalenessResponse": {
            "type": "object",
            "properties": {
                "age_minutes": {
                    "type": "integer",
                    "example": 15
                },
                "last_update": {
                    "type": "string",
                    "example": "2025-11-12T10:00:00Z"
                },
                "refresh_allowed": {
                    "type": "boolean",
                    "example": true
                },
                "region_id": {
                    "type": "integer",
                    "example": 10000002
                },
                "region_name": {
                    "type": "string",
                    "example": "The Forge"
                },
                "status": {
                    "description": "fresh, stale, very_stale",
                    "type": "string",
                    "example": "fresh"
                }
            }
        },
        "MarketOrderResponse": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "integer",
                    "example": 90
                },
                "is_buy_order": {
                    "type": "boolean",
                    "example": false
                },
                "issued": {
                    "type": "string",
                    "example": "2025-11-12T10:00:00Z"
                },
                "location_id": {
                    "type": "integer",
                    "example": 60003760
                },
                "min_volume": {
                    "type": "integer",
                    "example": 1
                },
                "order_id": {
                    "type": "integer",
                    "example": 123456789
                },
                "price": {
                    "type": "number",
                    "example": 5.5
                },
                "range": {
                    "type": "string",
                    "example": "region"
                },
                "system_id": {
                    "type": "integer",
                    "example": 30000142
                },
                "type_id": {
                    "type": "integer",
                    "example": 34
                },
                "volume_remain": {
                    "type": "integer",
                    "example": 100000
                },
                "volume_total": {
                    "type": "integer",
                    "example": 100000
                }
            }
        },
        "RegionResponse": {
            "type": "object",
            "properties": {
                "region_id": {
                    "type": "integer",
                    "example": 10000002
                },
                "region_name": {
                    "type": "string",
                    "example": "The Forge"
                }
            }
        },
...
        "TypeResponse": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 4
                },
                "group_id": {
                    "type": "integer",
                    "example": 18
                },
                "packaged_volume": {
                    "type": "number",
                    "example": 0.01
                },
                "type_id": {
                    "type": "integer",
                    "example": 34
                },
                "type_name": {
                    "type": "string",
                    "example": "Tritanium"
                },
                "volume": {
                    "type": "number",
                    "example": 0.01
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "EVE SSO Bearer Token. Format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}